hashmove updates

+++ hashmove.py
@@ -12,7 +12,7 @@ from optparse import OptionParser #OpenCube uses python 2.6 so has to be compati
 import getpass
 
 #figure out if arguments given are files or dirs, generate list of files and destinations to work on
-def makelist(startObj,dest,flist=[],soisdir=''):
+def makelist(startObj,dest,flist=[],soisdir=None):
 	if os.path.isfile(startObj):#if first argument is a file it's p easy
 		soisdir = False
 		endObj = os.path.join(dest, os.path.basename(startObj)) # this is the file that we wanted to move, in its destination
@@ -50,42 +50,44 @@ def hashfile(afile, hasher, blocksize=65536):
 	return
 
 #open the hash files and compare their contents
-def compareDelete(bar):
-	so, eo = bar #unpack tuples into startObject and endObject
-	soMD = so + ".md5" #add md5 extensions
-	eoMD = eo + ".md5"
+def compare(sfhashfile, efhashfile):
 	#open both md5 files
-	with open(soMD,'r') as f1:
-		with open(eoMD,'r') as f2:
-			sf1 = re.search('\w{32}',f1.read()) #this searches for a 32char alphanumeric string (the md5 hash)
-			ef2 = re.search('\w{32}',f2.read())
-			print "srce " + os.path.basename(so) + " " + sf1.group().lower()
-			print "dest " + os.path.basename(eo) + " " + ef2.group().lower()
-			if sf1.group().lower() == ef2.group().lower(): #if they're the same that's great
-				delyn = 1 # set an error level
+	with open(sfhashfile,'r') as f1:
+		with open(efhashfile,'r') as f2:
+			sfhash = re.search('\w{32}',f1.read()) #this searches for a 32char alphanumeric string (the md5 hash)
+			efhash = re.search('\w{32}',f2.read())
+			print "srce " + os.path.basename(sfhashfile) + " " + sfhash.group().lower()
+			print "dest " + os.path.basename(efhashfile) + " " + efhash.group().lower()
+			if sfhash.group().lower() == efhash.group().lower(): #if they're the same that's great
+				delyn = True # set an error level
 			else: #fine too but it means the files aren't the same on both ends of transfer
-				delyn = 0		
-	if delyn == 1: #if the strings match it's ok to delete these start file and its hash	
-		f1.close()
-		os.remove(so)
-		os.remove(soMD)
-	else:
-		print "uh there was an issue there"	
-	return
+				delyn = False
+	return delyn
+
 
 def main():
 	#initialize all of the things
 	usage = "Usage: python %prog [-options] source destination"
 	parser = OptionParser(usage=usage) #crate a parser object
-	parser.add_option('-c','--copy',action='store_true',dest='c',default='None',help="copy, don't delete from source") #add an option to our parser object, in this case '-c' for copy
-	parser.add_option('-l','--lto',action='store_true',dest='lto',default='None',help="write to lto, asks for lto barcode and logs files written")
+	parser.add_option('-c','--copy',action='store_true',dest='c',default=False,help="copy, don't delete from source") #add an option to our parser object, in this case '-c' for copy
+	parser.add_option('-l','--lto',action='store_true',dest='lto',default=False,help="write to lto, asks for lto barcode and logs files written")
+	parser.add_option('-q','--quiet',action='store_true',dest='q',default=False,help="quiet mode, don't print anything to console")
+	parser.add_option('-v','--verify',action='store_true',dest='v',default=False,help="verify mode, verifies sidecar hash(es) for file or dir")
 	(options, args) = parser.parse_args() #parse the parser object, return list of options followed by positional parameters
+	
 	#args[] is the index of a positional argument, 0=source, 1=destination, in this case
 	startObj = args[0].replace("\\","/") # fun fact, windows lets you type both fwd and back slashes in pathnames
-	dest = args[1].replace("\\","/") # just easier to replace these slashes for windows sry
-	if not os.path.exists(dest): # make the destination dir if it don't already exist
-		os.makedirs(dest)
-		
+	
+	if options.v is True: #if we're verifying then the start and end objects are the same
+		dest = startObj
+	else:
+		dest = args[1].replace("\\","/")
+	
+	#quiet mode redirects standard out to nul
+	if options.q is True:
+		f = open(os.devnull,'w')
+		sys.stdout = f
+	
 	#ok
 	#return a list of tuples, files to move and their destinations
 	flist, soisdir = makelist(startObj, dest)
@@ -109,29 +111,36 @@ def main():
 		print ""
 		print "hashing source and destination " + os.path.basename(sf)
 		#to change the hashing algorithm just replace MD5 with SHA256 or whatever, remember to change extensions up top too!
-		sf, hashfile(open(sf, 'rb'), hashlib.md5()), ef, hashfile(open(ef, 'rb'), hashlib.md5())
+		sf, hashfile(open(sf, 'rb'), hashlib.md5())
+		if options.v is False:
+			ef, hashfile(open(ef, 'rb'), hashlib.md5())
 	
-	
-	if options.c is 'None': #if we are moving not copying, as declared by flag -c, delete the start object
-		#compare and delete them
-		for f in flist:
+	#compare start and end hashes
+	for sf, ef in flist:
+		print ""
+		print "verifying source and destination hashes..."
+		sfhash = sf + ".md5"
+		efhash = ef + ".md5"
+		delyn = compare(sfhash,efhash)
+		if options.c is False and delyn is True:
 			print ""
-			print "verifying source and destination hashes..."
-			compareDelete(f)
+			print "deleting verified files..."
+			os.remove(sf)
+			os.remove(sfhash)
 			
-		#if we hashmoved a dir, try to delete the now empty dir
-		if soisdir is True:
-			rmDirList = []
-			for dirs, subdirs, files, in os.walk(startObj):
-				for x in subdirs: #delete subdirs
-					rmDirList.append(os.path.join(dirs,x))
-					#os.rmdir(os.path.join(dirs,x))
-			for rm in reversed(rmDirList):
-				os.rmdir(rm)
-				time.sleep(1.0) #gotta give the system time to catch up and recognize if a dir is empty
-			os.rmdir(startObj)
+	#if we hashmoved a dir, try to delete the now empty dir
+	if options.c is False and soisdir is True:
+		rmDirList = []
+		for dirs, subdirs, files, in os.walk(startObj):
+			for x in subdirs: #delete subdirs
+				rmDirList.append(os.path.join(dirs,x))
+				#os.rmdir(os.path.join(dirs,x))
+		for rm in reversed(rmDirList):
+			os.rmdir(rm)
+			time.sleep(1.0) #gotta give the system time to catch up and recognize if a dir is empty
+		os.rmdir(startObj)
 	
-	if options.lto is not 'None':
+	if options.lto is True:
 		ltoNumber = raw_input("What is the barcode of this LTO? ")
 		usr = getpass.getuser()
 		if usr == 'opencube':
@@ -148,6 +157,5 @@ def main():
 		ltoLog.write("\n")
 		ltoLog.close()
 	return
-
 
 main()
